Adds a navbar to the bottom of the page
The navbar-footer will be where the current team lives. Currently,
each button will take the user to the pokemons route, but the plan is
to have each individual pokemon occupy a button. When that button is
clicked, the pokemon page for that pokemon will come up and
modifications can be made.