Adds delete team member
Delete removes the Pokemon from your team. This is throwing an error
that I can not resolve, but the app is still functional. The error is
"Attempted to handle event `pushedData` on *model* while in state
root.deleted.inFlight." I'll continue to search for a solution, but as
of now, it's not breaking anything.